requirements.txt
flask 
pacy 
io
json
os
cv2
requests
csv